refactor: LottoNumbersFactory static 클래스인 점에 대한 코드리뷰 반영
RandomNumberGenerateStrategy 인터페이스 추가
LottoNumbersFactory -> FixedPoolRandomNumberGenerateStrategy로 static 클래스가 아닌 클래스로 변경
RandomNumberGenerateStrategy에 대한 테스트코드용 구현체 작성 후 테스트
