+++ src/main/java/lotto/domain/strategy/FixedPoolRandomNumberGenerateStrategy.java
@@ -1,19 +1,21 @@
-package lotto.domain;
+package lotto.domain.strategy;
+
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoNumbers;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class LottoNumbersFactory {
+public class FixedPoolRandomNumberGenerateStrategy implements RandomNumberGenerateStrategy {
 
   private static final List<LottoNumber> lottoNumberPool = IntStream.rangeClosed(LottoNumber.LOTTO_MIN_NUM, LottoNumber.LOTTO_MAX_NUM)
       .mapToObj(LottoNumber::new)
       .collect(Collectors.toList());
 
-  private LottoNumbersFactory() {}
-
-  public static LottoNumbers generateNewNumbers() {
+  @Override
+  public LottoNumbers generateNewNumbers() {
     Collections.shuffle(lottoNumberPool);
     List<LottoNumber> selectedNumbers = lottoNumberPool.subList(0, LottoNumbers.NUMBER_COUNT_PER_GAME);
     return new LottoNumbers(selectedNumbers);
