+++ src/main/java/lotto/domain/strategy/RandomNumberGenerateStrategy.java
@@ -1,0 +1,8 @@
+package lotto.domain.strategy;
+
+import lotto.domain.LottoNumbers;
+
+public interface RandomNumberGenerateStrategy {
+
+  LottoNumbers generateNewNumbers();
+}
