+++ src/test/java/lotto/domain/strategy/TestRandomNumberGenerateStrategy.java
@@ -1,0 +1,12 @@
+package lotto.domain.strategy;
+
+import lotto.domain.LottoNumbers;
+import lotto.util.TestUtil;
+
+public class TestRandomNumberGenerateStrategy implements RandomNumberGenerateStrategy {
+
+  @Override
+  public LottoNumbers generateNewNumbers() {
+    return new LottoNumbers(TestUtil.getOneToSixLottoNumbers());
+  }
+}
